fix(ci): preserve release quality ratchets
