+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
-  "maximum_collected_cases": 15765,
-  "maximum_test_source_lines": 340451,
+  "maximum_collected_cases": 15781,
+  "maximum_test_source_lines": 340722,
   "schema_version": 1
 }
