+++ tests/fixtures/guard-command-corpus/decision-diff-report.json
@@ -238,7 +238,7 @@
       "source-b6742d8f2f1ac4b54f46e427": "403f6f88f159e6257c04d8ed5726e62ba176cdee9abe04d60b8d0b0d0e3007b2",
       "source-b6d8d4f6c494653d8ac3b90c": "d03decff191f7abfa177fb166496f6ae9ceef34ae0079f6a4c56b4d64f041353",
       "source-b6ee2209355294e94b5580be": "3f2d0512ac75a5366725965aec5174628b5b05f732b00fa156a63b37cf278d88",
-      "source-b72fac2a9933d9c25c221a38": "bde743a523a9ebd42a04f970ecdc0dc38e086f86b71aab7a266db91cbf202d6a",
+      "source-b72fac2a9933d9c25c221a38": "57f301863062ef67209bdbd6585f34893c04497829f6aab9df9320115430717a",
       "source-b80a8083541fd7e8874926a6": "39c5cdd4522dbee6a92a8b20ee444d8b2b3d61c24de002d5ccdda129d5f8542c",
       "source-b933bc50f755d1a32c112e35": "5727d8a9e0f1bf39aa4470f048e346a2eed5b9437bd74c3c63e3420722e3b959",
       "source-b970eabd271372ddfe730bf1": "4446d801098f8517d245904b61eb4ba2992f1b7b25faccf283ee0fe15a2e07a0",
